fixed LP64 segfault in LSV
Recent pulls to my_types.h involved a change to the HALFWORD_ADDR_XOR mask to an unsigned constant (i.e., 02u instead of 02 signed int).  LSV in particular did multiplication of HES(0x000) which translated to (0 ^ 2u) which is not the same as (0 ^ 2).  Effectively, 64-bit systems did 0x00000000FFFFFFFFul * correction instead of 0xFFFFFFFFFFFFFFFFul * correction, which causes immediate crashing when (correction == 0) on valid implementations of the LP64 ABI.