Segregate bit-wise SP_STATUS_REG sets after clears.
This obfuscates the readable documentation of the SP_STATUS_REG bit resolutions in exchange for making it obvious how the bit patterns could be optimized (i.e., do all clears in one static operation and all sets in the next).  As said optimization would result in further obfuscation in exchange for micro-optimizing an infrequent scalar unit system control command, said optimization will not be done in this file.