Fix VCH op
Changed implementation of VCH opcode to be closer to how it is described in the docs, which now passes all tests of this test rom -> https://github.com/PeterLemon/N64/tree/master/RSPTest/CP2/VCL

However this implementation might not be entirely accurate/complete, for one the corner case hack of the original implementation is now not accounted for (not sure in what cases it would apply).
And in order to pass the third test for VCE it was necessary to compare with 2's complement instead of negative value (see comment in code), further test cases with real hardware might be useful here.