refined optimization from bf7c98f to account for very high dividends
Fixes #19.

Disabling the optimized code is perhaps a temporary measure, but the more readable code under the #else clause should absolutely be kept.  The optimized version for 2's complement machines has however also been patched with a fix in case it becomes desirable to go back to enabling it for substantial speed gains.