NDEBUG
Default to bypassing request for user acknowledgement of some
potentially fatal RSP messages.  This is not particularly wise, but
on some systems (for which we have not implemented drawing a dialog
box to shove the error on-screen to make it more apparent to the
user), it is annoying to find the console stream to do the keystroke
necessary to continue the emulation thread.  In general, this macro
will remain undefined for most non-Intel architectures.

Resolves #32.